Moved reEvaluation helper function and variables in LaunchRule+Reevaluable file
